CAMEL-11238 Run tests of JMS component against ...
...Apache Artemis

This is a work in progress it might not compile and tests might fail,
creating this for visibility and feedback.

This is an effort to test against Apache ActiveMQ Artemis the next
version of Active MQ but also to have a test harness in Camel JMS
component that can be used to run the tests against multiple queue
managers simultaneously.

The effort consists of modifying tests to be either run as integration
tests using a `MultipleJmsImplementations` JUnit Runner or putting mocks
in place where integration tests do not touch the JMS implementation
specifics -- i.e. converting some of the integration tests to unit
tests.