fix: skip executing integration tests in scr ar...
...chetype

When invoking via `maven-invoker` the MAVE_OPTS from the parent Maven
process are used, this leads to running the JVM with the same options as
the parent JVM process that is running Maven. So with generous heap size
this leads to Hotspot failing to allocate memory. The fix would need to
be done in the `maven-archetype-plugin` to set the `MAVEN_OPTS` of the
via `InvocationRequest::setMavenOpts` to a much more conservative memory
setting. Skipping this test and the execution of this second JVM seems
much more productive for now than waiting for months to get a PR merged.